+++ speech/src/main/java/com/example/speech/CreateRecognizerV2.java
@@ -35,6 +35,7 @@ public class CreateRecognizerV2 {
 
   public static void createRecognizerV2(String projectId, String recognizerId) throws IOException,
       InterruptedException, ExecutionException {
+
     // Initialize client that will be used to send requests. This client only needs to be created
     // once, and can be reused for multiple requests. After completing all of your requests, call
     // the "close" method on the client to safely clean up any remaining background resources.
