+++ speech/src/main/java/com/example/speech/TranscribeFileV2.java
@@ -42,8 +42,10 @@ public class TranscribeFileV2 {
 
   public static void transcribeFileV2(String recognizerName, String audioFilePath)
       throws IOException {
-    
-    // Instantiate the client
+
+    // Initialize client that will be used to send requests. This client only needs to be created
+    // once, and can be reused for multiple requests. After completing all of your requests, call
+    // the "close" method on the client to safely clean up any remaining background resources.
     try (SpeechClient speechClient = SpeechClient.create()) {
       Path path = Paths.get(audioFilePath);
       byte[] data = Files.readAllBytes(path);
