+++ speech/src/test/java/com/example/speech/QuickstartSampleV2IT.java
@@ -19,11 +19,10 @@ package com.example.speech;
 import static com.google.common.truth.Truth.assertThat;
 
 import com.google.api.gax.longrunning.OperationFuture;
-import com.google.cloud.ServiceOptions;
-import com.google.cloud.speech.v2.CreateRecognizerRequest;
 import com.google.cloud.speech.v2.DeleteRecognizerRequest;
 import com.google.cloud.speech.v2.OperationMetadata;
 import com.google.cloud.speech.v2.Recognizer;
+import com.google.cloud.speech.v2.RecognizerName;
 import com.google.cloud.speech.v2.SpeechClient;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
@@ -38,50 +37,31 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
+/** Tests for quickstart sample. */
 @RunWith(JUnit4.class)
 @SuppressWarnings("checkstyle:abbreviationaswordinname")
-public class TranscribeFileV2Test {
+public class QuickstartSampleV2IT {
   private String recognitionAudioFile = "./resources/commercial_mono.wav";
   private String recognizerId = String.format("rec-%s", UUID.randomUUID());
-  private String recognizerName;
   private String projectId = System.getenv("GOOGLE_CLOUD_PROJECT");
   private ByteArrayOutputStream bout;
   private PrintStream out;
+  private PrintStream originalPrintStream;
 
   @Before
-  public void setUp() throws IOException, ExecutionException, InterruptedException,
-      TimeoutException {
+  public void setUp() {
     bout = new ByteArrayOutputStream();
     out = new PrintStream(bout);
+    originalPrintStream = System.out;
     System.setOut(out);
-
-    // Create a recognizer for this test.
-    try (SpeechClient speechClient = SpeechClient.create()) {
-      String parent = String.format("projects/%s/locations/global", projectId);
-
-      // First, create a recognizer
-      Recognizer recognizer = Recognizer.newBuilder()
-          .setModel("latest_long")
-          .addLanguageCodes("en-US")
-          .build();
-
-      CreateRecognizerRequest createRecognizerRequest = CreateRecognizerRequest.newBuilder()
-          .setParent(parent)
-          .setRecognizerId(recognizerId)
-          .setRecognizer(recognizer)
-          .build();
-
-      OperationFuture<Recognizer, OperationMetadata> op =
-          speechClient.createRecognizerAsync(createRecognizerRequest);
-      recognizer = op.get(180, TimeUnit.SECONDS);
-      recognizerName = recognizer.getName();
-    }
   }
 
   @After
-  public void tearDown() throws IOException, ExecutionException, InterruptedException,
+  public void tearDown() throws IOException, InterruptedException, ExecutionException,
       TimeoutException {
-    System.setOut(null);
+    System.setOut(originalPrintStream);
+
+    String recognizerName = RecognizerName.format(projectId, "global", recognizerId);
 
     DeleteRecognizerRequest deleteRequest = DeleteRecognizerRequest.newBuilder()
         .setName(recognizerName)
@@ -97,7 +77,7 @@ public class TranscribeFileV2Test {
   @Test
   public void testQuickstart() throws Exception {
     // Act
-    TranscribeFileV2.transcribeFileV2(recognizerName, recognitionAudioFile);
+    QuickstartSampleV2.quickstartSampleV2(projectId, recognitionAudioFile, recognizerId);
 
     // Assert
     String got = bout.toString();
